requirements.txt
discord.py==1.3.3
python-3.8.0
pytz==2019.3